Remove requirement for IDs
This removes the requirement to have unique numerical IDs per project, by relying upon the unique filenames for the `.json` files within `/lib/project` instead.

This is a good idea, I think, as it removes the potential for accidental ID conflicts when different people write different Pull Requests to add projects at the same time (or for people to accidentally forget to add an ID altogether, as I did).  By relying on the filenames we can ensure that they are always present and always unique.

The downside to this is that it will break all the existing project URLs, which contain the IDs and a "slugified" version of the project name (which may or may not match the json filename).