Update design and layout
This moves the layout away from being based in columns for service status (alpha, beta, live and unknown), as these are increasingly irrelevant or hard to ascertain. Instead the services are shown as a simple list.

The navigation has also moved out of the black header section and into a grey bar, for improved visibility.

The font has switched to Helvetica Neue instead of GDS Transport, as this is not an official GOV.UK service.