Add homepage with masthead (#925)
Now that [the site](https://govuk-digital-services.herokuapp.com) is
using the service navigation component, this creates a new homepage for
it which uses the [Masthead
component](https://x-govuk.github.io/govuk-prototype-components/masthead/)
underneath it.

This also has the advantage of making the 'By topic' page more visible
in the nav, and giving a bit more space to explain the project.

I've also tried moving the 'latest milestones' list from the right-hand
column of the list to the main homepage, so it's a bit more visible on
mobile. Not sure how useful that 'updates' list is, but it's something
to indicate a bit of activity and hopefully tempt people to update the
data with new service changes.

Thoughts?

Possibly the service navigation doesn't need the 1px solid dark-grey
border at the bottom of it when the masthead is immediately beneath it?

### Screenshot

<img width="1231" alt="Screenshot 2024-09-09 at 23 20 24"
src="https://github.com/user-attachments/assets/1d2e1dc5-bb35-4f88-b562-474bd7d965d3">