Update head.html (#995)
Linked to the favicon in the `<head>` of the site. Browsers usually find
it automatically if in the root directory, but as it isn't located
there, it isn't being detected and doesn't work in my latest Chrome
version.

Example below:
<img width="646" alt="Screenshot 2024-12-17 at 09 24 41"
src="https://github.com/user-attachments/assets/34636241-35d8-4f6c-a87d-c166aae6af8b"
/>

Co-authored-by: Frankie Roberto <frankie@frankieroberto.com>